+++ mealie/routes/organizers/controller_tags.py
@@ -12,7 +12,7 @@ from mealie.schema.recipe.recipe_category import TagSave
 from mealie.schema.response.pagination import PaginationQuery
 from mealie.services import urls
 from mealie.services.event_bus_service.event_bus_service import EventBusService
-from mealie.services.event_bus_service.message_types import EventOperation, EventTagData, EventTypes
+from mealie.services.event_bus_service.event_types import EventOperation, EventTagData, EventTypes
 
 router = APIRouter(prefix="/tags", tags=["Organizer: Tags"])
